+++ src/async_std.rs
@@ -2,6 +2,7 @@ use
 {
 	crate        :: { SpawnHandle, LocalSpawnHandle, JoinHandle, join_handle::InnerJh } ,
 	futures_task :: { FutureObj, LocalFutureObj, Spawn, LocalSpawn, SpawnError        } ,
+	futures_util :: { future::abortable                                               } ,
 	std          :: { sync::atomic::AtomicBool                                        } ,
 };
 
@@ -76,10 +77,13 @@ impl<Out: 'static + Send> SpawnHandle<Out> for AsyncStd
 {
 	fn spawn_handle_obj( &self, future: FutureObj<'static, Out> ) -> Result<JoinHandle<Out>, SpawnError>
 	{
+		let (fut, a_handle) = abortable( future );
+
 		Ok( JoinHandle{ inner: crate::join_handle::InnerJh::AsyncStd
 		{
-			handle  : Some( async_std_crate::task::spawn( future ) ) ,
-			detached: AtomicBool::new( false )                       ,
+			handle  : async_std_crate::task::spawn( fut ) ,
+			detached: AtomicBool::new( false )            ,
+			a_handle                                      ,
 		}})
 	}
 }
@@ -96,9 +100,9 @@ impl<Out: 'static + Send> SpawnHandle<Out> for AsyncStd
 
 		Ok( JoinHandle{ inner: InnerJh::AsyncStd
 		{
-			handle  : Some( async_std_crate::task::spawn( future ) ) ,
-			detached: AtomicBool::new( false )                       ,
-			a_handle                                                 ,
+			handle  : async_std_crate::task::spawn_local( fut ) ,
+			detached: AtomicBool::new( false )                  ,
+			a_handle                                            ,
 		}})
 	}
 }
@@ -109,10 +113,13 @@ impl<Out: 'static> LocalSpawnHandle<Out> for AsyncStd
 {
 	fn spawn_handle_local_obj( &self, future: LocalFutureObj<'static, Out> ) -> Result<JoinHandle<Out>, SpawnError>
 	{
+		let (fut, a_handle) = abortable( future );
+
 		Ok( JoinHandle{ inner: InnerJh::AsyncStd
 		{
-			handle  : Some( async_std_crate::task::spawn_local( future ) ) ,
-			detached: AtomicBool::new( false )                             ,
+			handle  : async_std_crate::task::spawn_local( fut ) ,
+			detached: AtomicBool::new( false )            ,
+			a_handle                                      ,
 		}})
 	}
 }